refactor: match detekt rule
